change training and crawling method

+++ train/train.py
@@ -2,7 +2,7 @@ import pandas as pd
 from tqdm import tqdm
 import pickle
 from model import model
-from config import dataset_dir,dataset_file_name, dataset_link_path,dataset_file_exp,model_dir, model_file_name, model_file_exp, model_link_path
+from config import dataset_dir,dataset_file_name, dataset_link_path, model_dir, model_file_name, model_link_path
 from crawling import *
 import os
 import time
@@ -11,6 +11,7 @@ def train_model(model_name : str , cfg : dict):
     if model_name == 'EASE' :
 
         ## 유저-문제 크롤링 세팅 ##
+        print(f"유저 아이디 크롤링...")
         user_ids = []
         try:
             page_st = cfg['page_st']
@@ -36,6 +37,7 @@ def train_model(model_name : str , cfg : dict):
                 add_group_ids(group_num, num, user_ids)
 
         ## 유저-문제 크롤링 ##
+        print(f"유저-문제 크롤링...")
         start = time.time()
 
         user_problem = {}
@@ -56,17 +58,20 @@ def train_model(model_name : str , cfg : dict):
                 user_problem_mat[user][problem] = 1
 
         ## dataframe으로 저장 - 바로가기 생성 ##
-        os.unlink(dataset_link_path)
+        if os.path.islink(dataset_link_path):
+            os.unlink(dataset_link_path)
 
         df = pd.DataFrame(user_problem_mat).T
+        dataset_file_exp = 'csv'
         dataset_path = os.path.join(dataset_dir,f"{dataset_file_name}_{end}.{dataset_file_exp}")
         df.to_csv(dataset_path)
 
-        os.symlink(dataset_path, model_link_path)
+        os.symlink(dataset_path, dataset_link_path)
 
         df = df.to_numpy()
 
         ## EASE 훈련 ##
+        print(f"EASE 훈련...")
         start = time.time()
 
         ease = model.EASE(300)
@@ -79,12 +84,16 @@ def train_model(model_name : str , cfg : dict):
         # 훈련은 약 10초 걸림 - 유저 500명 기준 - 시립대 학생들이 푼 문제만
 
         # 모델 저장 - 바로가기 생성
+        model_file_exp = 'p'
         model_path = os.path.join(model_dir,f'{model_file_name}_{end}.{model_file_exp}')
         with open(model_path, 'wb') as file:  # 파일을 바이너리 쓰기 모드(wb)로 열기
-            os.unlink(model_link_path)
+            if os.path.islink(model_link_path):
+                os.unlink(model_link_path)
 
             pickle.dump(ease, file)
 
             os.symlink(model_path, model_link_path)
 
-    return '훈련 완료'
+    print('훈련 완료')
+
+    return
